# java-client/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
  /api/v1/geoCache:
    post:
      description: Send a geo cache submission to the endpoint.
      operationId: sendSubmission
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoCacheSubmission'
        required: true
      responses:
        "200":
          description: OK
      tags:
      - geo-cache
      x-content-type: application/json
      x-accepts: application/json
  /api/v1/geoCache/{pageNumber}:
    get:
      description: Retrieve geo cache submission.
      operationId: getSubmission
      parameters:
      - explode: false
        in: path
        name: pageNumber
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GeoCacheSubmission'
                type: array
          description: OK
      tags:
      - geo-cache
      x-accepts: application/json
components:
  schemas:
    GeoCacheSubmission:
      example:
        date: date
        note: note
        latitude: latitude
        name: name
        secret: secret
        longitude: longitude
      properties:
        name:
          maxLength: 50
          minLength: 3
          pattern: "^[ -~]+$"
          type: string
        note:
          maxLength: 250
          minLength: 3
          pattern: "^[ -~]+$"
          type: string
        secret:
          maxLength: 50
          minLength: 3
          pattern: "^[ -~]+$"
          type: string
        longitude:
          maxLength: 20
          minLength: 1
          pattern: "^[ -~]+$"
          type: string
        latitude:
          maxLength: 20
          minLength: 1
          pattern: "^[ -~]+$"
          type: string
        date:
          type: string
      required:
      - latitude
      - longitude
      - name
      - note
      - secret
      type: object